Drastically simplify our macros
This honestly was just silly.

I don't think we need to be calling in different macros from different
files. This was a call out previously by others, but, was too much work
for the time.

Eat the frog.

Need to determine if anything breaks, but I think everything should be
fine. BCBIO should not be required, but, I left it in there _just in
case_ we have something in the tools that is calling it. Otherwise,
everything else is pretty much the vanilla setup for our tools.